change color
太さの所はあれから変えておりません
