+++ js/draw.js
@@ -37,19 +37,9 @@ define(function(require) {
             this.drag = false;
             context.beginPath();
         },
-        
+
     };
        
-    var colors = document.getElementById('colors').childNodes;
-    for (var i = 0, color; color = colors[i]; i++) {
-        if (color.nodeName.toLowerCase() != 'div') continue;
-        color.addEventListener('click', function (e) {
-            var style = e.target.getAttribute('style');
-            var color = style.match(/background:(#......)/)[1];
-            context.strokeStyle = color;
-            context.fillStyle = color;
-        },false);
-    };
 
     var sizes= document.getElementById('sizes').childNodes;
     for (var i = 0, size; size = sizes[i]; i++){
@@ -57,6 +47,7 @@ define(function(require) {
         size.addEventListener('click', function(e){
             var size = e.currentTarget.getAttribute('data-size');
             context.lineWidth = size;
+            radius = size;
         }, false);
     };
